ai-103-agent-multi/requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0
azure-identity>=1.17.0
azure-monitor-opentelemetry>=1.6.0
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
